+++ R/weightFunctions.R
@@ -27,11 +27,11 @@
 #' @export
 #'
 #' @examples
-#' slopePenaltyFn(0.5,0.51)
-#' slopePenaltyFn(0.5,0.65)
+#' gradePenaltyFn(0.5,0.51)
+#' gradePenaltyFn(0.5,0.65)
 #' # grade > 20% so NA
-#' slopePenaltyFn(0.5,0.75)
-slopePenaltyFn<-function(x1,x2,resolution=1, baseCost = 16178,limit=20,penalty=504,limitWeight=NA){
+#' gradePenaltyFn(0.5,0.75)
+gradePenaltyFn<-function(x1,x2,resolution=1, baseCost = 16178,limit=20,penalty=504,limitWeight=NA){
   #If one of the nodes is a road or barrier ignore grade penalty
   cond = pmin(x1,x2)>=0
   cond[is.na(cond)]=F