+++ tests/testthat/test-AltGraphFns.R
@@ -80,21 +80,21 @@
 #   ext_2000 <- ext(c(base_point, base_point+200*the_res)[c(1,3,2,4)])
 #   landscape_2000 <- list(weightRaster=crop(landscape, ext_2000))
 # 
-#   rr = getGraph(landscape_100,neighb,method="old",weightFunction=slopePenaltyFn,limitWeight=65000)
+#   rr = getGraph(landscape_100,neighb,method="old",weightFunction=gradePenaltyFn,limitWeight=65000)
 # 
 #   neighb  ="octagon"
 #   bm <- bench::mark(min_iterations = 1, check = FALSE,
 #                     old_100 = getGraph(landscape_100,neighb,method="old"),
-#                     dem_100 = getGraph(landscape_100,neighb,method="old",weightFunction = slopePenaltyFn),
+#                     dem_100 = getGraph(landscape_100,neighb,method="old",weightFunction = gradePenaltyFn),
 #                     gdistance_100 = getGraph(landscape_100,neighb,method="gdistance"),
 #                     old_500 = getGraph(landscape_500,neighb,method="old"),
-#                     dem_500 = getGraph(landscape_500,neighb,method="old",weightFunction=slopePenaltyFn),
+#                     dem_500 = getGraph(landscape_500,neighb,method="old",weightFunction=gradePenaltyFn),
 #                     gdistance_500 = getGraph(landscape_500,neighb,method="gdistance"),
 #                     old_1000 = getGraph(landscape_1000,neighb,method="old"),
-#                     dem_1000 = getGraph(landscape_1000,neighb,method="old",weightFunction=slopePenaltyFn),
+#                     dem_1000 = getGraph(landscape_1000,neighb,method="old",weightFunction=gradePenaltyFn),
 #                     gdistance_1000 = getGraph(landscape_1000,neighb,method="gdistance"),
 #                     old_2000 = getGraph(landscape_2000,neighb,method="old"),
-#                     dem_2000 = getGraph(landscape_2000,neighb,method="old",weightFunction=slopePenaltyFn),
+#                     dem_2000 = getGraph(landscape_2000,neighb,method="old",weightFunction=gradePenaltyFn),
 #                     gdistance_2000 = getGraph(landscape_2000,neighb,method="gdistance")
 #   )
 #