+++ NEWS.md
@@ -1,7 +1,7 @@
 # roads 1.1.2
 * Change dynamic least cost paths (DLCP) to iterative least cost paths (ILCP) throughout
 * Change default `roadMethod` to `"ilcp"` in `projectRoads` 
-* Add ability to use a custom `weightFunction` and add a `weightFunction` `slopePenaltyFn` that determines the grade between two cells
+* Add ability to use a custom `weightFunction` and add a `weightFunction` `gradePenaltyFn` that determines the grade between two cells
 * Change argument name from `cost` to `weightRaster` since it no longer represents a cost surface
 and can now be inputs to the `weightFunction`. Also change `roadsInCost` to `roadsInWeight`.
 * returned roads are no longer unioned together.